dags/requirements.txt
snowflake-connector-python==2.7.1
snowflake-sqlalchemy==1.4.2
watchtower==1.0.6
airflow-dbt==0.4.0
dbt-core==1.2.2
dbt-postgres==1.2.2
psycopg2-binary==2.9.3
